added Layout component to wrap home and other pages
